fixed jumping too far past boundary of cell error
We had the following bug: the ray is near the boundary of the cell, and jumps forward into the next cell by a distance equal to the localSDF it can see in this cell. But this is too far for something in the next cell, and so we teleport inside of it. To fix, we binary search around the cell boundary to get ourselves outside the cell, but not much outside. We do binary search rather than raytracing the cell boundary, because in other geometries we can't ray trace!