Change the model of points in the shader
Previously points where vec4 in the model H^2 x R
This cause many numerical errors, since going back and forth between this representation and the actual element of SO(2,1) involved arctan

Now points are a pair (vec4, float) where the vec4 component describes an element of SL(2,R) and the float the fiber coordinates.

In this way we correct the (wrong) behavior of bouncing balls.